Eliminate AutoConfig
It really sucks, but nobody is using this, and the front-end interface is going to be tough to clean up.  Eliminate it, all tests, and all associated support functionality.

Nevertheless, this did motivate a replacement concept that has syntax to LeapSerial branch.  This is implemented in `feature-config`.
