+++ src/autowiring/autowiring_error.cpp
@@ -1,9 +1,17 @@
 // Copyright (C) 2012-2015 Leap Motion, Inc. All rights reserved.
 #include "stdafx.h"
 #include "autowiring_error.h"
+#include "demangle.h"
+#include <sstream>
 
 autowiring_error::autowiring_error(const std::string& what):
   m_what(what)
 {}
 
 autowiring_error::~autowiring_error(void){}
+
+void autowiring::ThrowFailedTypeParseException(const std::string& str, const std::type_info& ti) {
+  std::stringstream msg;
+  msg << "Failed to parse '" << str << "' as type '" << autowiring::demangle(ti) << "'";
+  throw autowiring_error(msg.str());
+}
